Remove the option to override existing dimensions. Change dimension init phase.
Unregistering dimensions has the problem that it is impossible to
unregister the DimensionType. Thus trying to override an existing
registered dimension could lead to using the wrong DimensionType in some
cases, although most things should get it from the World. But if something
tries to get the DimensionType from DimensionType.getById(), it would
return the old one most likely.