Add a GamemodeTracker
The tracker keeps track of the player's "normal" gamemode in
non-forced-gamemode dimensions. It also changes the gamemode for players
to the one set for a dimension, if the ForceGamemode option is also true
for that dimension. When the player returns from any forced-gamemode
dimension to a normal non-forced-gamemode dimension, the previously
stored gamemode is returned for them and removed from the tracker.