Fix player data being broken if using WorldInfo override in dimension 0
Also fixed not re-reading the per-dimension WorldInfo after the world save
directory is known, and also removed setting the player dimension in the
NBT just for the crashreports, because that would actually modify the live
player data (unless the tag would be copied...).

Also we don't save the level.dat for dimension 0, but instead we let
vanilla/Forge handle it in that case.