Refactored where the colors tag is stored
The colors tag is now stored directly inside the dimension entry, instead
of nested inside the worldinfo section. This way using WorldInfo overrides
is not needed for color overrides to work.

Also, the color values are not stored per-dimension on the client anymore.
Instead they are synced and set every time the player changes dimensions.
There is also a flag indicating if they are not used, to try to optimize
the biome color event execution time a bit.