Add a 'require_exact_match' options for the 'dimensiontype' object
Currently it require the properties of the WorldProvider class,
shouldLoadSpawn, dimtypeId and suffix to match.

The only difference to the 'allow_different_id' option is thus the
requirement for the suffix string to also match. The name is not checked,
because it makes very little difference to anything.

The shouldLoadSpawn and WorldProvider class values must always
match even in the most leanient configuration.