Fix storing the wrong "non-forced" gamemode on initial spawn
The player's current dimension was not stored on initial login, so the
player login event then mistakenly thought that the initial join was a
dimension change, and stored/overrode the "normal gamemode" with the
gamemode that the player had in the initial spawn dimension.

Also some other minor clean-up related to the DataTracker and initial
spawn handling code.